+++ doc/changes.rst
@@ -9,6 +9,8 @@ Version 2019.6
     This version is currently under development.
 
 * :func:`Indexer` renamed to :func:`~pystella.index_fields`.
+* :meth:`Sector.get_args` deprecated in favor of using
+  :func:`~pystella.get_field_args`.
 * Various improvements and bug fixes.
 
 Version 2019.5